Refactor: extract scheduled parsing logic from SheetsParser and config fix
- Moved data extraction logic to dedicated SheetsParserService
- Enabled Async
- SheetsParser now delegates scheduled execution via AsyncParserService
- Simplified SheetsParser to only trigger periodic updates
- Added graceful shutdown of ExecutorService using finally block
- Made thread pool size configurable via config
- Fixed config name in SheetNamesConfig
- Added ConfigurationPropertiesScan annotation on main for reading from yml c
