+++ src/main/resources/application.properties
@@ -9,4 +9,3 @@ spring.web.resources.chain.enabled=true
 server.servlet.encoding.charset=UTF-8
 server.servlet.encoding.enabled=true
 server.servlet.encoding.force=true
-
